Change kmer_size to type u64 to avoid conversions in QueryFileParser

+++ src/QueryFileParser/QueryFileParser.h
@@ -19,16 +19,15 @@ namespace sbwt_search {
 class QueryFileParser {
 private:
   string filename;
-  uint kmer_size;
+  u64 kmer_size;
   bool has_parsed = false;
   void check_if_has_parsed();
   void add_sequence(const string &seq);
   vector<string> seqs;
   u64 total_letters = 0;
   u64 total_positions = 0;
-
 public:
-  QueryFileParser(const string &filename, const uint kmer_size):
+  QueryFileParser(const string &filename, const u64 kmer_size):
     filename(filename), kmer_size(kmer_size){};
   auto &get_seqs() { return seqs; };
   auto get_total_letters() { return total_letters; };
